requirements.txt
flask
flask_cors
selenium
python-dotenv
